refactor: rename propsWithTypesPipe to typedPropsPipe
